Move util function to utils.js
