+++ frontend/src/App.vue
@@ -136,7 +136,7 @@
 
 <script>
 import { config, routingContractConfig } from "./contracts/contracts-config";
-import { getScanHtml } from './utils/utils.js'
+import { getScanHtml, hexStringToArrayBuffer } from './utils/utils.js'
 import TxState from "./enums/tx-state";
 import ErrorType from './enums/error-type'
 import Web3 from "web3";
@@ -479,7 +479,7 @@ export default {
             this.routingContract.address,
             // this.fromCfxToDrip(this.csv.vals.reduce((a, b) => a + b, 0)),
             sum.toString(),
-            this.hexStringToArrayBuffer(data)
+            hexStringToArrayBuffer(data)
           );
 
           const estimate = await tx.estimateGasAndCollateral({
@@ -496,7 +496,7 @@ export default {
         } else {
           const tx = this.routingContract.send(
             // this.fromCfxToDrip(this.csv.vals.reduce((a, b) => a + b, 0)),
-            this.hexStringToArrayBuffer(data)
+            hexStringToArrayBuffer(data)
           );
 
           const estimate = await tx.estimateGasAndCollateral({
@@ -592,38 +592,6 @@ export default {
           // this.errorMessage = err.message;
       }
       // console.log(this.errors)
-    },
-    
-    // util function
-    hexStringToArrayBuffer(hexString) {
-      // remove the leading 0x
-      hexString = hexString.replace(/^0x/, "");
-
-      // ensure even number of characters
-      if (hexString.length % 2 != 0) {
-        console.log(
-          "WARNING: expecting an even number of characters in the hexString"
-        );
-      }
-
-      // check for some non-hex characters
-      var bad = hexString.match(/[G-Z\s]/i);
-      if (bad) {
-        console.log("WARNING: found non-hex characters", bad);
-      }
-
-      // split the string into pairs of octets
-      var pairs = hexString.match(/[\dA-F]{2}/gi);
-
-      // convert the octets to integers
-      var integers = pairs.map(function(s) {
-        return parseInt(s, 16);
-      });
-
-      var array = new Uint8Array(integers);
-      // console.log(array);
-
-      return array.buffer;
     },
     setCsv(csv) {
       this.csv = csv
